# "==" can be redefined