```
     ARY = Array.new(100) { |i| i }
      class Integer
        def == other
          ARY.replace([])
          self.equal?(other)
        end
      end
      ARY.count(42) == 0

=> true
```


unfortunatelly isMethodBuiltin is very expensive
```
require 'benchmark/ips'
a = Array.new(200, 6)

Benchmark.ips do |x|
  x.report('include') { a.include?(5) }
end
```

this patch
```
 244.569k (± 4.4%) i/s -      1.230M in   5.042204s
```
indy (without num-opts)
```
 288.265k (± 6.7%) i/s -      1.444M in   5.036346s
```
master
```
 835.808k (± 9.6%) i/s -      4.134M in   4.998483s
```
mri
```
 681.590k (± 9.3%) i/s -      3.389M in   5.020591s
```