@kares 2018-10-03
hey, is this a real use-case you run into or just trying to fix the test?

maybe this can get redesigned - assuming objects of same type a call-site might help.
in which case, ideally, there should be a single isBuiltin check against the (cached) site ...