@ahorek 2018-10-03
I just want to increase the test coverage. This fix is a corner case,  but I think there're other places where redefinition should be considered and is_buildin method shouldn't be that slow. I'll try something with indy.

This shouldn't be merged until the performance problem is solved even if the behaviour is correct.