@ahorek 2018-10-11
I'm not sure if I'm using it correctly. Now it's better, but still 2 times slower than before. Any idea how to improve it?