@kares 2018-10-12
looks reasonably good, do you happen to know which part is slowest, maybe the `isRefinement` check?