@enebo 2018-10-12
@ahorek another idea if it is isBuiltin taking the time would be to save builtin EQ method somewhere similar to what we do in context.sites and then compare the method in callsite against that saved version.  That would simplify this to a single equality check; but also would require you do save it at some point in time during bootstrapping of Ruby.