@enebo 2024-05-21
in 2022 There was some very similar code landed for this method (it is missing the !isRefinement part).  Closing just due to age and sorry we did not come to consensus on this.