@ahorek 2024-05-21
you're right, it has already been fixed by https://github.com/jruby/jruby/commit/505dea7dc141d9d95b899bbb95d737441b16c2b9 thanks for noticing!